backend/requirements.txt
fastapi
uvicorn
firecrawl-py
groq
python-dotenv
requests
pydantic
streamlit
httpx
textstat
